requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.10.0

# GUI dependencies (optional - for Streamlit app)
streamlit>=1.28.0
plotly>=5.18.0

# Development dependencies (optional)
pytest>=7.0
pytest-cov>=4.0
